Latest loxi

+++ src/python/loxi/generic_util.py
@@ -11,6 +11,7 @@ Utility functions independent of the protocol version
 
 import loxi
 import struct
+import functools
 
 def pack_list(values):
     return "".join([x.pack() for x in values])
@@ -82,7 +83,7 @@ class OFReader(object):
         self.offset += length
 
     def skip_align(self):
-        new_offset = (self.offset + 7) / 8 * 8
+        new_offset = int((self.offset + 7) / 8) * 8
         if new_offset > self.length:
             raise loxi.ProtocolError("Buffer too short")
         self.offset = new_offset
